pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agmem"
version = "0.1.0"
description = "Agentic Memory Version Control System - Git for AI agent memories"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "agmem Team", email = "team@agmem.dev"}
]
keywords = ["ai", "agent", "memory", "version-control", "git", "vcs", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]
mcp = [
    "mcp>=1.0.0",
]
vector = [
    "sqlite-vec>=0.1.0",
    "sentence-transformers>=2.0.0",
]
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
]
cloud = [
    "boto3>=1.34.0",
    "google-cloud-storage>=2.0.0",
]
gardener = [
    "scikit-learn>=1.3.0",
    "openai>=1.0.0",
]
pii = [
    "presidio-analyzer>=2.2.0",
]
daemon = [
    "watchdog>=3.0.0",
]
graph = [
    "networkx>=3.0",
]
all = [
    "mcp>=1.0.0",
    "sqlite-vec>=0.1.0",
    "sentence-transformers>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "boto3>=1.34.0",
    "google-cloud-storage>=2.0.0",
    "scikit-learn>=1.3.0",
    "openai>=1.0.0",
    "watchdog>=3.0.0",
    "networkx>=3.0",
]

[project.scripts]
agmem = "memvcs.cli:main"

[project.urls]
Homepage = "https://github.com/vivek-tiwari-vt/agmem"
Documentation = "https://github.com/vivek-tiwari-vt/agmem#readme"
Repository = "https://github.com/vivek-tiwari-vt/agmem"
"Bug Tracker" = "https://github.com/vivek-tiwari-vt/agmem/issues"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
